+++ nautilus/src/nautilus/tracing/TraceOperation.hpp
@@ -47,8 +47,8 @@ struct IndirectFunctionCall {
 };
 
 struct BlockRef {
-	explicit BlockRef(uint16_t block);
-	uint16_t block;
+	explicit BlockRef(uint32_t block);
+	uint32_t block;
 	std::vector<TypedValueRef> arguments;
 };
