+++ nautilus/src/nautilus/tracing/phases/SSACreationPhase.hpp
@@ -83,7 +83,7 @@ private:
 		 * @brief Returns the set of value refs defined by operations in the given block,
 		 * computing and caching it on first access.
 		 */
-		const std::unordered_set<ValueRef>& getOrBuildDefinitions(uint16_t blockId);
+		const std::unordered_set<ValueRef>& getOrBuildDefinitions(uint32_t blockId);
 
 		/**
 		 * @brief Hoists all ALLOCA operations to the head of the initial block.
@@ -109,12 +109,12 @@ private:
 		std::unordered_set<uint32_t> processedBlocks;
 		// Lazy cache: maps blockId → set of value refs produced by operations in that block.
 		// Entries are built on first access rather than upfront.
-		std::unordered_map<uint16_t, std::unordered_set<ValueRef>> blockDefinitions;
+		std::unordered_map<uint32_t, std::unordered_set<ValueRef>> blockDefinitions;
 		// Tracks (blockId, ref) pairs that have already been propagated,
 		// replacing the O(n) std::find on predBlock.arguments.
 		std::unordered_set<uint64_t> propagatedValues;
 		// Reused across propagateValue calls to avoid repeated heap allocation.
-		std::vector<uint16_t> propWorklist;
+		std::vector<uint32_t> propWorklist;
 	};
 };
